src/requirements.txt
fastapi==0.115.12
pydantic==2.11.1
SQLAlchemy==2.0.40
uvicorn==0.34.0
asyncpg==0.30.0
sqlalchemy_searchable==2.1.0
alembic==1.15.1
openpyxl==3.1.5
python-jose==3.4.0
python_multipart==0.0.20
dotenv==0.9.9
passlib==1.7.4
bcrypt==3.2.2
cryptography==44.0.2
httpx==0.28.1
aiogram==3.22.0
pillow==11.3.0